+++ docs/metro/js/metro.js
@@ -1,5 +1,5 @@
 /*
- * Metro 4 Components Library v4.2.25 build 703 (https://metroui.org.ua)
+ * Metro 4 Components Library v4.2.26 build @@build (https://metroui.org.ua)
  * Copyright 2018 Sergey Pimenov
  * Licensed under MIT
  */
@@ -92,8 +92,8 @@ var isTouch = (('ontouchstart' in window) || (navigator.MaxTouchPoints > 0) || (
 
 var Metro = {
 
-    version: "4.2.25",
-    versionFull: "4.2.25.703 ",
+    version: "@@version",
+    versionFull: "@@version.@@build @@status",
     isTouchable: isTouch,
     fullScreenEnabled: document.fullscreenEnabled,
     sheet: null,
@@ -3615,8 +3615,8 @@ var d = new Date().getTime();
     getCursorPosition: function(el, e){
         var a = Utils.rect(el);
         return {
-            x: e.pageX - a.left - window.pageXOffset,
-            y: e.pageY - a.top - window.pageYOffset
+            x: Utils.pageXY(e).x - a.left - window.pageXOffset,
+            y: Utils.pageXY(e).y - a.top - window.pageYOffset
         };
     },
 
@@ -10700,10 +10700,9 @@ var ImageCompare = {
         var overlay = element.find(".image-container-overlay");
         var slider = element.find(".image-slider");
 
-        element.on(Metro.events.start, ".image-slider", function(e){
+        slider.on(Metro.events.start, function(e){
             var w = element.width();
-            e.preventDefault();
-            $(window).on(Metro.events.move + "-" + element.attr("id"), function(e){
+            $(document).on(Metro.events.move + "-" + element.attr("id"), function(e){
                 var x = Utils.getCursorPositionX(element, e), left_pos;
                 if (x < 0) x = 0;
                 if (x > w) x = w;
@@ -10716,9 +10715,9 @@ var ImageCompare = {
                 });
                 Utils.exec(o.onSliderMove, [x, left_pos, slider[0]], element[0]);
             });
-            $(window).on(Metro.events.stop + "-" + element.attr("id"), function(){
-                $(window).off(Metro.events.move + "-" + element.attr("id"));
-                $(window).off(Metro.events.stop + "-" + element.attr("id"));
+            $(document).on(Metro.events.stop + "-" + element.attr("id"), function(){
+                $(document).off(Metro.events.move + "-" + element.attr("id"));
+                $(document).off(Metro.events.stop + "-" + element.attr("id"));
             })
         });
 
@@ -10923,8 +10922,6 @@ var ImageMagnifier = {
         if (o.magnifierMode !== "glass") {
             cx = zoomElement[0].offsetWidth / glass_size / 2;
             cy = zoomElement[0].offsetHeight / glass_size / 2;
-
-            console.log(cx, cy);
 
             zoomElement.css({
                 backgroundSize: (image.width * cx) + "px " + (image.height * cy) + "px"